+++ src/locale/BasisTexts.ts
@@ -243,6 +243,8 @@ type BasisTexts = {
             /** See `en-US.json` for documentation  */
             random: FunctionText<[]>;
             /** See `en-US.json` for documentation  */
+            shuffled: FunctionText<[]>;
+            /** See `en-US.json` for documentation  */
             first: FunctionText<[]>;
             /** See `en-US.json` for documentation  */
             last: FunctionText<[]>;