+++ src/locale/OutputTexts.ts
@@ -55,6 +55,17 @@ type OutputTexts = {
         /** $1 = Layout description, $2 = pose description */
         description: Template;
     } & TypeTexts;
+    /** A shadow */
+    Aura: NameAndDoc & {
+        /** The shadow's color */
+        color: NameAndDoc;
+        /** The horizontal offset of the shadow */
+        offsetX: NameAndDoc;
+        /** The vertical offset of the shadow */
+        offsetY: NameAndDoc;
+        /** The blurriness of the shadow */
+        blur: NameAndDoc;
+    };
     /** A sequence of glyphs */
     Phrase: NameAndDoc & {
         /** The glyphs to render */
@@ -67,6 +78,8 @@ type OutputTexts = {
         direction: NameAndDoc;
         /** The matter properties for the phrase */
         matter: NameAndDoc;
+        /** The shadow properties for the phrase */
+        aura: NameAndDoc;
         /** A description of the phrase for screen readers. 1$: non-optional text, $2: optional name, $3: optional size, $4: optional font, $5: then non-optional pose */
         description: Template;
     } & TypeTexts;
@@ -156,7 +169,7 @@ type OutputTexts = {
         scale: NameAndDoc;
         flipx: NameAndDoc;
         flipy: NameAndDoc;
-        /** all optional inputs: opacity, rotation, scale, flipx, flipy */
+        /** Templated description of the pose */
         description: Template;
     };
     /** A sequence of poses, keyed by percentage complete, for use in overriding an output's defaults for entering, resting, moving, or existing states */